Refactored composite binary file loading
