+++ mhdata/merge/binary/parsers/__init__.py
@@ -1,0 +1,2 @@
+from .structs_ext import *
+from .mib import QuestBinary, load_quest
